speed improvement by removing redundant compute
